The changes introduced in this fork use a C implementation of the scrypt/USDE merkle root hashing algorythm.
The 'standard bitcoin merkle tree hash method' is roughly as follows;
The coinbase "script signature" of the first transaction is hashed into a 32 byte segment. This segment is then hashed against each transaction that is associated with the block. The final result is filled into the final block data known as the "merkle root hash".

To elaborate further; the merkle root effectively ensures that all transactions are properly accounted for when the block is stored into the chain.

The USDE coin's protocol uses a more complicated method of calculating the merkle root hash from the associated transactions.

The following is a c++ psuedo-code sample of how the merkle root hash is calculated in the USDE (and other scrypt) daemons:
{
  foreach (CTransation tx, vtx) {
    vMerkleTree.push_back(tx.GetHash());
  }
  foreach (CTransation tx, vtx) {
      uint256 merkleHash = 0;
      vMerkleTree.push_back(tx.GetHash());
      int j = 0;
      for (int nSize = vtx.size(); nSize > 1; nSize = (nSize + 1) / 2)
      {
        for (int i = 0; i < nSize; i += 2)
        {
          int i2 = std::min(i+1, nSize-1);
          vMerkleTree.push_back(Hash(BEGIN(vMerkleTree[j+i]),  END(vMerkleTree[j+i]), BEGIN(vMerkleTree[j+i2]), END(vMerkleTree[j+i2])));
        }
        j += nSize;
      }
      if (!vMerkleTree.empty())
        merkleHash = vMerkleTree.back();
      return (merkleHash);
  }
}

If you print out the actual hashes are they are computed you may note the more complicated method of combining the transactions in hashed form. For example, the algorythm above will perform 3 hash operations, as opposed to 2 operations with traditional method, when there are 2 total transactions associated with the block.

What is misleading here is that when less than 2 transactions are associated with the block being mined the scrypt merkle root equation will equal that of the SHA256 bitcoin model. This means any time there are more than 2 transactions pending the merkle root computed, used to determine whether a nonce is valid enough to meet a target, will not actually reflect the merkle root used by the server to complete the block.